requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==2.1.1
Flask-Cors==3.0.10
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.4
openai==0.28.0
openpyxl==3.1.5
pandas==2.0.3
pandas-stubs==2.0.3.230814
propcache==0.2.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
types-pytz==2024.2.0.20241221
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.0.3
yarl==1.15.2
zipp==3.20.2
